Remove lightdm and instead autostart i3/sway from tty1
This was primarily due to a race condition that resulted in XF86 keys
not working in sway as a result of VT switching. The keys work when sway
is launched from tty1, so instead of using a display manager I'm just
launching it directly.

Note: The window manager that is started is controlled by the
$DISPLAY_MANAGER env variable in ~/.profile. Both the sway and i3 dirs
have their own version of .profile, so whichever dir is stowed last will
be started on login.